requirements.txt
meshtastic>=2.7.5
pyserial>=3.5
pynmea2>=1.19.0

